# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Multi-Agent Career Opportunity Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/resumes": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Resumes",
        "operationId": "list_resumes_resumes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumeResponse"
                  },
                  "type": "array",
                  "title": "Response List Resumes Resumes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Upload Resume",
        "operationId": "upload_resume_resumes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/resumes/{resume_id}": {
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Delete Resume",
        "operationId": "delete_resume_resumes__resume_id__delete",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumeProfile"
                  },
                  "type": "array",
                  "title": "Response List Profiles Profiles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/profiles/{profile_id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/outreach": {
      "get": {
        "tags": [
          "outreach"
        ],
        "summary": "List Outreach",
        "operationId": "list_outreach_outreach_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OutreachStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutreachResponse"
                  },
                  "title": "Response List Outreach Outreach Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/outreach/{outreach_id}": {
      "get": {
        "tags": [
          "outreach"
        ],
        "summary": "Get Outreach",
        "operationId": "get_outreach_outreach__outreach_id__get",
        "parameters": [
          {
            "name": "outreach_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outreach Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/outreach/{outreach_id}/approve": {
      "post": {
        "tags": [
          "outreach"
        ],
        "summary": "Approve Outreach",
        "operationId": "approve_outreach_outreach__outreach_id__approve_post",
        "parameters": [
          {
            "name": "outreach_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outreach Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveOutreachRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/outreach/{outreach_id}/reject": {
      "post": {
        "tags": [
          "outreach"
        ],
        "summary": "Reject Outreach",
        "operationId": "reject_outreach_outreach__outreach_id__reject_post",
        "parameters": [
          {
            "name": "outreach_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outreach Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/outreach/{outreach_id}/edit": {
      "post": {
        "tags": [
          "outreach"
        ],
        "summary": "Edit Outreach",
        "operationId": "edit_outreach_outreach__outreach_id__edit_post",
        "parameters": [
          {
            "name": "outreach_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outreach Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditOutreachRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveOutreachRequest": {
        "properties": {
          "final_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Message"
          }
        },
        "type": "object",
        "title": "ApproveOutreachRequest"
      },
...
      "CreateResumeRequest": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_content": {
            "type": "string",
            "title": "File Content"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "file_content"
        ],
        "title": "CreateResumeRequest"
      },
...
      "EditOutreachRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "EditOutreachRequest"
      },
      "EducationEntry": {
        "properties": {
          "institution": {
            "type": "string",
            "title": "Institution"
          },
          "degree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          }
        },
        "type": "object",
        "required": [
          "institution"
        ],
        "title": "EducationEntry"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutreachChannel": {
        "type": "string",
        "enum": [
          "EMAIL",
          "LINKEDIN_MESSAGE",
          "LINKEDIN_CONNECTION_REQUEST",
          "OTHER"
        ],
        "title": "OutreachChannel",
        "description": "Owner: Outreach Service."
      },
      "OutreachResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "contact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contact Id"
          },
          "channel": {
            "$ref": "#/components/schemas/OutreachChannel"
          },
          "draft_message": {
            "type": "string",
            "title": "Draft Message"
          },
          "final_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Message"
          },
          "status": {
            "$ref": "#/components/schemas/OutreachStatus"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "contact_id",
          "channel",
          "draft_message",
          "status",
          "generated_at"
        ],
        "title": "OutreachResponse"
      },
      "OutreachStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PENDING_APPROVAL",
          "APPROVED",
          "REJECTED",
          "EDITED",
          "SENT",
          "SEND_FAILED"
        ],
        "title": "OutreachStatus",
        "description": "Owner: Outreach Service. See state-machines.md#outreach-lifecycle-outreachstatus--outreachstatus."
      },
...
      "ProfileStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "ARCHIVED"
        ],
        "title": "ProfileStatus",
        "description": "Owner: Resume/Profile Service. See\nstate-machines.md#candidateprofile-lifecycle-candidateprofilestatus--profilestatus."
      },
...
      "ResumeProfile": {
        "properties": {
          "profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Profile Id"
          },
          "resume_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resume Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "seniority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationEntry"
            },
            "type": "array",
            "title": "Education",
            "default": []
          },
          "certifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certifications",
            "default": []
          },
          "projects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Projects",
            "default": []
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries",
            "default": []
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Roles",
            "default": []
          },
          "status": {
            "$ref": "#/components/schemas/ProfileStatus"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "resume_id",
          "user_id",
          "title",
          "status"
        ],
        "title": "ResumeProfile",
        "description": "The read/composition type combining a Resume and its derived\nCandidateProfile. Used whenever another component needs \"a user's\nprofile\" \u2014 most notably as an input to Job Matching.\n\nProduced by: Resume/Profile Service (GET /profiles, GET /profiles/{id}).\nConsumed by: Job Matching Service (as JobMatchingState.profiles).\nNot persisted directly \u2014 assembled on read from resumes +\ncandidate_profiles."
      },
      "ResumeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "status": {
            "$ref": "#/components/schemas/ResumeStatus"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "file_name",
          "status",
          "uploaded_at"
        ],
        "title": "ResumeResponse"
      },
      "ResumeStatus": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "PARSING",
          "PARSED",
          "PARSE_FAILED",
          "ARCHIVED"
        ],
        "title": "ResumeStatus",
        "description": "Owner: Resume/Profile Service. See state-machines.md#resume-lifecycle-resumestatus--resumestatus."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}